company-srv/requirements.txt
# backend/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.10.3
python-dotenv==1.0.1
coloredlogs==15.0.1
loguru==0.7.3
asyncpg==0.30.0
greenlet==3.1.1
python-jose[cryptography]==3.3.0 
passlib[bcrypt]==1.7.4